i9300: don't pull down the power key
On reset, GPX2-7 is set to pull-down. This causes the PMIC
to think that the power key is being pressed, and so causes the PMIC
to reset the SoC after ~8 seconds (as would happen if you held the power
key down normally).

To fix this, we do what S-BOOT does: set the input to floating, and rely
on the external pull-up to make the key work.