drivers: introduce the extcon subsystem
Loosely modelled on the Linux kernel's equivalent, the extcon subsystem
provides a generic layer for detecting whether or not cables are plugged
in, the type of said cables, and the ability of said cables to charge
the board.

The "extcon" subsystem loosely occupies the position of a "MUIC" in the
PMIC dm hierarchy.

Signed-off-by: Simon Shields <simon@lineageos.org>