i9300: env updates
Add documentation of env variables, and add commands to update from the
"bootloader" partition on the eMMC. Note that the command does some
(very) basic validation of the bootloader payload - it checks that
there's something resembling a secure boot payload located at the end of
the BL1 within the bootloader image. This won't prevent malicious
attacks, but it will prevent user error/installing clearly broken
bootloaders.